+++ CHANGES.md
@@ -1,10 +1,14 @@
 # Unreleased
 
+### Added
 - Support for OxCaml unboxed named types (@art-w, #1407)
 - Support for OxCaml zero alloc definitions (@Leonidas-from-XIV, #1422, #1444)
 - Remove requirement for ppx_expect in tests (@jonludlam, #1445)
 - Support for OxCaml modalities (@art-w, #1420)
 - Fix resolving functor through `module type of` (@Leonidas-from-XIV, #1471)
+
+### Fixed
+- Allow to break link into multiline (@Tim-ats-d, #1439)
 
 # 3.2.1