+++ CHANGES.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Support for OCaml 5.6 (@jonludlam, #1472)
 - Support for OxCaml unboxed named types (@art-w, #1407)
 - Support for OxCaml zero alloc definitions (@Leonidas-from-XIV, #1422, #1444)
 - Remove requirement for ppx_expect in tests (@jonludlam, #1445)